feat: change data type Long to Interger

+++ src/main/java/floud/demo/dto/friendship/FriendshipListResponseDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Builder
 public class FriendshipListResponseDto {
     private String my_nickname;
-    private Long totalFriendNum;
-    private Long totalPage;
+    private Integer totalFriendNum;
+    private Integer totalPage;
     private List<FriendshipDto> friendshipList;
 }
